+++ lib/magic/presenter/base.rb
@@ -1,0 +1,25 @@
+# frozen_string_literal: true
+
+require 'delegate'
+
+module Magic
+	module Presenter
+		# = Magic Presenter
+		#
+		# Any presenter should inherit `Magic::Presenter::Base`.
+		# It further inherits from `SimpleDelegator` and behaves like that.
+		#
+		# == Presenter class inference
+		#
+		# Presenters provide automatic class inference for any model based
+		# on its class name powered by Magic Lookup.
+		# For example, `MyModel.new` looks for `MyModelPresenter` first.
+		# If not found, it looks for presenters of its ancestor classes,
+		# up to `ObjectPresenter`.
+		class Base < Decorator::Base
+			class << self
+				def name_for(object_class) = "#{object_class}Presenter"
+			end
+		end
+	end
+end