+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 ## [0.1.0] — UNRELEASED
 
+### Added
+
+- `Magic::Presenter::Base` — a basic presenter class.